requirements.txt
accelerate==1.1.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.2
numpy==2.1.3
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
psutil==6.1.0
pyclipper==1.3.0.post6
pytesseract==0.3.13
python-bidi==0.6.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
sympy==1.13.1
tifffile==2024.9.20
tokenizers==0.15.2
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
verovio==4.3.1
Werkzeug==3.1.3